Separated module and Pic A fitting
