+++ src/SyncSSCModel.jl
@@ -1,116 +1,202 @@
 module SyncSSCModel
 
 using Parameters
-using LaTeXStrings
-using HCubature
-using Plots
-gr()   # default backend for Plots
+using PhysicalConstants
+using Unitful
+using UnitfulAstro
+using SpectralFitting
+# using Plots
 
-# synchrotron.jl
-export dn_e
-export j_syn
-export S_syn
-export P_syn
-export syncSpec
-export syncPlot
+import PhysicalConstants.CODATA2018: c_0, σ_e, m_e, h
 
-# Compton.jl
-export j_ssc
-export P_ssc
-export comptonSpec
-export ssc_Plot
+qconvert(T::Type, q::Unitful.Quantity) = convert(T, q.val)
 
-# Synchrotron_General.jl
-export j_nu
-export S_nu
-export Sync_Plot
+const h_cgs_f = qconvert(Float64, uconvert(u"erg*s", h))
+const m_e_cgs_f = qconvert(Float64, uconvert(u"g", m_e))
+const c_cgs_f = qconvert(Float64, uconvert(u"cm/s", c_0))
+const σ_e_cgs_f = qconvert(Float64, uconvert(u"cm^2", σ_e))
 
-export MyParamStruct
+# module Callmodel
 
-# Parameters and constants in cgs units
-@with_kw struct MyParamStruct
-    # set the default type of all parameters
-    @deftype Float64
-
-    # Physical constants
-    "Mass of electron in g"
-    m_e = 9.1093897E-28
-    "Charge of electron in esu - electrostatic system units"
-    e = 4.8032068E-10
-    "Speed of light in cm s^-1"
-    c = 2.99792458E10
-    "Planck's constant in ergs"
-    h = 6.6260755E-27
-    "Thomson cross section in cm^2"
-    σ_T = 0.66524616E-24
-    "Permeability of free space in G cm-1/2 g-1/2 s2"
-    mu_o = 4.191696447656766E-10
-    
-    # Parameters specific to the physics of the problem
-    "Magnetic field strength in Gauss"
-    B = 1.5E-5 # PKS0637-752
-    #B = 2.8E-6 # Pictor A  
-    #B = [1E-6, 1E-4, 1E-2] # Just to test for a list of B-values
-    # PKS0637-752: Values taken from Schwartz et al. 2000 or Tavecchio et al. 2000 or Uchiyama et al. 2005
-    "Normalisation of electron density in cm^-3"
-    n_e0 = 6.0E-5 # # PKS0637-752
-    #n_e0 = 485 # Pictor A
-    #n_e0 = [1E-5, 10, 450] # Just to test for a list of n_e0-values
-    "Power law index of electron distribution function"
-    p = 2.7 # PKS0637-752
-    #p = 3.3 # Pictor A
-    #p = [2, 2.5, 3] # Just to test for a list of B-values            
-    "Minimum Lorentz factor of electrons"
-    γ_min = 1.78E3  # PKS0637-752 # Calculated using the expression in section 3.1 Uchiyama et al. 2005
-    #γ_min = 1.0E2 # Pictor A
-    "Maximum Lorentz factor of electrons"
-    γ_max = 3.21E5  # PKS0637-752 # Calculated using the expression in section 3.1 Uchiyama et al. 2005
-    #γ_max = 4E6 # Pictor A
-    "Redshift"
-    z = 0.651 # PKS0637-752
-    #z = 0.035 # Pictor A
-    "Luminosity distance in cm"
-    dL = 1.26E28
-    "Bulk Lorentz factor"
-    Γ = 12.0 # PKS0637-752
-    #Γ = 1 # Pictor A
-    #Γ = [1, 5, 10] # Just to test for a list of Γ-values
-    "Angle (radians) between the direction of the blob's motion and the direction to the observer"
-    θ = 3.5*pi/180.0 # PKS0637-752
-    #θ = 23*pi/180.0 # Pictor A
-    #θ = [0, 5, 10] # Just to test for a list of θ-values
-    "Radius of emitting region"
-    radius = 3.086E21 # PKS0637-752
-    #radius = 7.7E20 # Pictor A
-    "Mass of BH in Solar Masses"
-    M_BH = 1.0E8
-
-    "Hubble parameter"
-    ho = 0.67
-    "Hubble constant in km s^-1 Mpc^-1 in cgs units the Hubble constant should be in the units of s^-1"
-    # Hubble Constant = (50-100) km s^-1 Mpc^-1 = (1.6-3.2) 10^-18 s^-1
-    # Ho = 100.0*ho
-    Ho = 71.0 # PKS0637-752
-    #Ho = 50 # Pictor A
-
-    # Calculated constants needed needed for calculations
-    # I think we should move these to a function rather than have them as "parameters" because they could get out of sync
-    "Cyclotron energy in units of m_e*c^2"
-    ϵ_B = B/4.414E13    # Dermer 1995 below eq.8
-    "Magnetic field energy density"
-    u_B = B^2/(8.0*pi)
-
-    # Binning for problem
-    # log_10 low frequency 
-    log_ν_low = 7.0
-    # log_10 high frequency
-    log_ν_high = 26.0
-    # number of logarithmic frequency bins
-    ν_n::Int = 100
+# Minor functions
+"""
+Cyclotron energy in units of m_e*c^2
+"""
+function ϵ_B(mps)
+    return(mps.B/4.414E13)
 end
 
-include("synchrotron.jl")
-include("Compton.jl")
-include("Synchrotron_General.jl")
+"""
+Magnetic field energy density
+"""
+function u_B(mps)
+    return((mps.B)^2/(8.0*pi))
+end
 
-end # module
+"""
+Lorentz factor
+"""
+function γ(ϵ, mps)
+    return(sqrt(ϵ / ϵ_B(mps)))
+end
+
+"""
+Spectral Index / power law index
+"""
+function α(mps)
+    return((mps.p - 1.0) / 2.0)
+end
+
+"""
+Doppler factor
+"""
+function δ(mps)
+    β = sqrt(1.0 - 1.0/mps.Γ^2)
+    μ_obs = cos(mps.θ)
+    return(1.0 / (mps.Γ*(1 - μ_obs*β)))
+end
+
+"""
+Volume Sphere
+"""
+function Vb(mps)
+    return((4.0/3.0) * pi * mps.radius^3)
+end
+
+# Major functions
+"""
+Electron density distribution
+"""
+function dn_e(ϵ, mps)
+    if (γ(ϵ, mps) < mps.γ_min) || (γ(ϵ, mps) > mps.γ_max)
+        dn_e = 0.0
+    else
+        dn_e = mps.n_e0 * γ(ϵ, mps)^-mps.p
+    end
+    return dn_e
+end
+"""
+Synchrotron Emissivity
+"""
+function j_syn(ϵ, mps)
+    return(((c_cgs_f*σ_e_cgs_f*u_B(mps))/(6.0*pi*ϵ_B(mps))) * γ(ϵ, mps) * dn_e(ϵ, mps))
+end
+"""
+Synchrotron Flux density
+"""
+function S_syn(ϵ, mps)
+    return((δ(mps)^3 * (1.0+mps.z) * Vb(mps) * j_syn(ϵ*(1.0+mps.z)/δ(mps), mps)) / mps.dL^2)
+end
+"""
+Synchrotron Spectral Power Flux
+"""
+function syncSpec(log_ν, mps)
+    ν = 10.0.^log_ν
+    ϵ = h_cgs_f*ν/(m_e_cgs_f*c_cgs_f^2)
+    spec = zeros(length(log_ν))
+    for ix in range(1, length(log_ν))
+        spec[ix] = S_syn(ϵ[ix], mps)
+    end
+    return(spec)
+end
+"""
+Synchrotron Self-Compton emissivity
+"""
+function j_ssc(ϵ, mps)
+    Σ_c = min(ϵ^-1, ϵ/mps.γ_min^2, ϵ_B(mps)*mps.γ_max^2) / max(ϵ_B(mps)*mps.γ_min^2, ϵ/mps.γ_max^2)
+    if Σ_c > 1.0
+        return(((c_cgs_f * σ_e_cgs_f^2 * mps.n_e0^2 * u_B(mps) * mps.radius) / (9.0 * pi * ϵ_B(mps))) * (ϵ/ϵ_B(mps))^-α(mps) * log(Σ_c))
+    else
+        return(0.0)
+    end
+end
+"""
+Synchrotron Self-Compton Flux density
+"""
+function P_ssc(ϵ, mps)
+
+    Σ_c = min(δ(mps)/(ϵ*(1.0+mps.z)), mps.γ_max^2 * ϵ_B(mps), ϵ*(1.0+mps.z)/(δ(mps)*mps.γ_min^2)) / max(mps.γ_min^2 * ϵ_B(mps), ϵ*(1+mps.z)/(δ(mps)*mps.γ_max^2))
+    if Σ_c > 1.0
+        return(((δ(mps))^(3.0+α(mps)) * (c_cgs_f*σ_e_cgs_f^2*mps.n_e0^2*u_B(mps)*mps.radius*Vb(mps)) * (1.0+mps.z)^(1.0-α(mps)) * (ϵ/ϵ_B(mps))^(1.0-α(mps)) * log(Σ_c)) / (9.0*pi*mps.dL^2))
+    else
+        return(0.0)
+    end
+end
+"""
+Synchrotron Self-Compton Spectral Power Flux
+"""
+function comptonSpec(log_ν, mps)
+    ν = 10.0.^log_ν
+    ϵ = h_cgs_f*ν/(m_e_cgs_f*c_cgs_f^2)
+    spec = zeros(length(log_ν))
+    for ix in range(1, length(log_ν))
+        spec[ix] = P_ssc(ϵ[ix], mps) / ϵ[ix]
+    end
+    return(spec)
+end
+
+# the export keyword only really works when 
+# developing with packages and modules are pre-compiled
+# export S_syn, P_ssc
+# end
+
+# Main programme for the Sync+SSC Model
+# using .Callmodel
+
+# need mean function
+# using StatsBase   # Commented this, not yet needed for this part of fitting
+
+function evaluate_model(ν, mps)
+    ϵ = h_cgs_f*ν/(m_e_cgs_f*c_cgs_f^2)
+
+    # Get synchrotron
+    syncFluxDensity = S_syn(ϵ, mps)
+    # Get self-Compton
+    comptonFluxDensity = P_ssc(ϵ, mps) / ϵ
+    return (syncFluxDensity + comptonFluxDensity) * ϵ
+end
+
+function evaluate_model(νs::AbstractArray, mps)
+    map(ν -> evaluate_model(ν, mps), νs)
+end
+
+@with_kw struct SSCModel{T,F} <: AbstractSpectralModel{T,Additive}
+    @deftype T
+    # need this field for additive models
+    "Normalisation."
+    K::T
+    "Magnetic field strength in Gauss"
+    B::T
+    "Normalisation of electron density in cm^-3"
+    n_e0::T
+    "Power law index"
+    p::T
+    "Minimum Lorentz Factor"
+    γ_min::T
+    "Maximum Lorentz Factor"
+    γ_max::T
+    "Bulk Lorentz Factor"
+    Γ::T
+    "Radius of emitting Region"
+    radius::T
+    "Angle (radians) between the direction of the blob's motion and the direction to the observer"
+    θ::T
+    "Luminosity distance"
+    dL::T
+    "Redshift"
+    z::T
+end
+
+# ensure input length is same as output length
+SpectralFitting.Δoutput_length(::Type{<:SSCModel}) = 0
+
+# specify how the model should be invoked
+# note: passing ν not log_10(ν)
+# note: returning flux not log_10(flux) 
+@inline function SpectralFitting.invoke!(flux, ν, model::SSCModel)
+    @. flux = evaluate_model(ν, (model,))
+end
+
+export SSCModel
+
+end
